refactor: reorganize bank composables by removing index and types files, and updating imports for better modularity
