+++ app/src/composables/bod/functions.ts
@@ -2,7 +2,7 @@ import { encodeFunctionData, type Address } from 'viem'
 import { useToastStore, useTeamStore } from '@/stores'
 import { useBodWrites } from './writes'
 import { BOD_FUNCTION_NAMES } from './types'
-import { useValidation } from '@/composables/bank'
+import { useValidation } from '@/composables/bank/utils'
 import type { Action } from '@/types'
 import { computed, ref, watch } from 'vue'
 import { estimateGas, readContract, writeContract } from '@wagmi/core'
